feat: add Tiptap bubble menu
